[TorchFix] add visitor for deprecated TorchVision transform (#4615)
This adds a violation when the
[deprecated](https://github.com/pytorch/vision/blob/ddfee23d56700ba84fd28805d5cbdeac2f28f2a7/torchvision/transforms/v2/_deprecated.py#L12-L13)
`torchvision.transforms.v2.ToTensor` transform is imported or used.

For the future, there are two more things to do:

1. The corresponding functional is also
[deprecated](https://github.com/pytorch/vision/blob/ddfee23d56700ba84fd28805d5cbdeac2f28f2a7/torchvision/transforms/v2/functional/_deprecated.py#L10-L11)
and needs the same treatment as the function. If we keep the long
violation message (which is the same message that one would see using
this at runtime), we probably should have a different error code for the
functional to avoid making the message even longer. Otherwise we may
also have both under the same error code.
2. Instead of just complaining about the use of deprecated
functionality, we could also provide a fix. However, this is not
straight forward, since we need to replace `v2.ToTensor()` with
`v2.Compose([v2.ToImage(), v2.ToDtype(torch.float32, scale=True)])`.
While that should work everywhere, usually `v2.ToTensor` is not used as
standalone, but rather already as part of a `Compose`, e.g.
    
    ```python
    transform = transforms.Compose([
        transforms.Resize(...),
        transforms.CenterCrop(...),
        transforms.ToTensor(),
        transforms.Normalize(...),
    ])
    ```
    
    or
    
    ```python
    pipeline = [
        transforms.Resize(...),
        transforms.CenterCrop(...),
    ]
    
    if foo:
        pipeline.append(...)
    else:
        pipeline.extend(...)
    
    pipeline.extend([
        transforms.ToTensor(),
        transforms.Normalize(...),
    ])
    transform = transforms.Compose(pipeline)
    ```

In both cases, we want don't want to have an extra `Compose` in there,
but just replace `ToTensor` with `v2.ToImage(),
v2.ToDtype(torch.float32, scale=True)`.

I don't have much experience with `libcst`, but from what I did in this
PR I think 1. is fairly straight forward, while 2. can become a
nightmare.