Fix deprecated symbol loading in zipped deployments
torchfix currentyl uses standard filesystem methods to find and load the
deprecated symbols data from disk. When running flake8 as zipped
deployments, this fails because the data is not accessible as a standard
filesystem path.

This replaces the filesystem usage with stdlib `pkgutil.get_data()` [1]
that is capable of resolving the data file within the zip deployment,
and loading that data using the correct internal mechanisms.

1: https://docs.python.org/3/library/pkgutil.html#pkgutil.get_data