torchfix: Refactor ERROR_CODE to be consistent
Refactors ERROR_CODE --> ERROR, introduces a new TorchError class that
defines what an error is including its code + the message it's supposed
to have since I found that most errors were defined by the Code at the
top + a string message that was semi-hard to find amongst the code.

This also makes it so that ERRORS needs to be a list instead of  being
optionally a list or a value to make it more consistent.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>