Improving test case developer experience (#69)
This PR updates the test fixtures to have the `.py` extension for
codemod output files. This is helpful for writing new tests (I'll be
adding some for the new rules)

- Codemod output is now stored as `.out.py` instead of `.py.out`. This
marks the files as Python source to IDEs, linters etc. For test
discovery, the input files were renamed to `.in.py` from `.py`.

Follow-up PR for https://github.com/pytorch-labs/torchfix/pull/67. Will
rebase this one once that PR is merged.